Mention alternative names for data driven tests in README
I was looking around for existing packages that make writing data driven tests in mocha more convenient to write. I like the API of this package but did not find it initially because I was searching using the terms _table driven_ and _parameterized_ tests. A quick Google search says that _data driven_ is the most common term but the two are also frequently used. Hopefully this will help more people find this.